delete unnecessary, add exception message, policy make it better, add middleware, change from authorize to middleware

See merge request energeek/universitas-pertamina-api!170

+++ routes/api.php
@@ -29,9 +29,9 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/', 'TitleController@index');
         Route::get('/select-list', 'TitleController@selectList');
         Route::get('/{title}', 'TitleController@show');
-        Route::post('/', 'TitleController@store');
-        Route::put('/{title}', 'TitleController@store');
-        Route::delete('/{title}', 'TitleController@destroy');
+        Route::post('/', 'TitleController@store')->middleware('can:admin-only');
+        Route::put('/{title}', 'TitleController@store')->middleware('can:admin-only');
+        Route::delete('/{title}', 'TitleController@destroy')->middleware('can:admin-only');
     });
 
     Route::get('/v2/title/', 'TitleController@getForDatatables');
@@ -40,18 +40,18 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/', 'NationalityController@index');
         Route::get('/select-list', 'NationalityController@selectList');
         Route::get('/{nationality}', 'NationalityController@show');
-        Route::post('/', 'NationalityController@store');
-        Route::put('/{nationality}', 'NationalityController@store');
-        Route::delete('/{nationality}', 'NationalityController@destroy');
+        Route::post('/', 'NationalityController@store')->middleware('can:admin-only');
+        Route::put('/{nationality}', 'NationalityController@store')->middleware('can:admin-only');
+        Route::delete('/{nationality}', 'NationalityController@destroy')->middleware('can:admin-only');
     });
 
     Route::group(['prefix' => 'skill'], function () {
         Route::get('/', 'SkillController@index');
         Route::get('/select-list', 'SkillController@selectList');
         Route::get('/{skill}', 'SkillController@show');
-        Route::post('/', 'SkillController@store');
-        Route::put('/{skill}', 'SkillController@store');
-        Route::delete('/{skill}', 'SkillController@destroy');
+        Route::post('/', 'SkillController@store')->middleware('can:admin-only');
+        Route::put('/{skill}', 'SkillController@store')->middleware('can:admin-only');
+        Route::delete('/{skill}', 'SkillController@destroy')->middleware('can:admin-only');
     });
 
     Route::group(['prefix' => 'department'], function () {
@@ -59,18 +59,18 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/select-list', 'DepartmentController@selectList');
         Route::get('/get-list-institution', 'DepartmentController@getListInstitution');
         Route::get('/{department}', 'DepartmentController@show');
-        Route::post('/', 'DepartmentController@store');
-        Route::put('/{department}', 'DepartmentController@store');
-        Route::delete('/{department}', 'DepartmentController@destroy');
+        Route::post('/', 'DepartmentController@store')->middleware('can:admin-only');
+        Route::put('/{department}', 'DepartmentController@store')->middleware('can:admin-only');
+        Route::delete('/{department}', 'DepartmentController@destroy')->middleware('can:admin-only');
     });
 
     Route::group(['prefix' => 'institution'], function () {
         Route::get('/', 'InstitutionController@index');
         Route::get('/select-list', 'InstitutionController@selectList');
         Route::get('/{institution}', 'InstitutionController@show');
-        Route::post('/', 'InstitutionController@store');
-        Route::put('/{institution}', 'InstitutionController@store');
-        Route::delete('/{institution}', 'InstitutionController@destroy');
+        Route::post('/', 'InstitutionController@store')->middleware('can:admin-only');
+        Route::put('/{institution}', 'InstitutionController@store')->middleware('can:admin-only');
+        Route::delete('/{institution}', 'InstitutionController@destroy')->middleware('can:admin-only');
     });
 
     Route::group(['prefix' => 'regulation'], function () {
@@ -89,11 +89,11 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/get-institution', 'OpportunityController@getInstitution');
         Route::get('/get-type-opportunity', 'OpportunityController@getTypeOpportunity');
         Route::get('/{opportunity}', 'OpportunityController@show');
-        Route::post('/', 'OpportunityController@store');
+        Route::post('/', 'OpportunityController@store')->middleware('can:create,App\Models\Opportunity');
         Route::post('/{opportunity}', 'OpportunityController@storeFiles');
         Route::post('/{opportunity}/interest', 'OpportunityController@interest');
-        Route::put('/{opportunity}', 'OpportunityController@store');
-        Route::delete('/{opportunity}', 'OpportunityController@destroy');
+        Route::put('/{opportunity}', 'OpportunityController@store')->middleware('can:create,opportunity');
+        Route::delete('/{opportunity}', 'OpportunityController@destroy')->middleware('can:delete,opportunity');
         Route::get('/files/{opportunityFile}', 'OpportunityController@showFile');
     });
 
@@ -128,12 +128,12 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
     });
 
     Route::group(['prefix' => 'role'], function () {
-        Route::get('/', 'RoleController@index');
+        Route::get('/', 'RoleController@index')->middleware('can:admin-only');
         Route::get('/select-list', 'RoleController@selectList');
-        Route::get('/{role}', 'RoleController@show');
-        Route::post('/', 'RoleController@store');
-        Route::put('/{role}', 'RoleController@store');
-        Route::delete('/{role}', 'RoleController@destroy');
+        Route::get('/{role}', 'RoleController@show')->middleware('can:admin-only');
+        Route::post('/', 'RoleController@store')->middleware('can:admin-only');
+        Route::put('/{role}', 'RoleController@store')->middleware('can:admin-only');
+        Route::delete('/{role}', 'RoleController@destroy')->middleware('can:admin-only');
     });
 
 
@@ -147,8 +147,8 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/accept-invitation', 'ResearchUserController@acceptInvitation');
         Route::post('/change-institution/{member}', 'ResearchUserController@changeInstitution');
         Route::post('/{member}', 'ResearchUserController@store');
-        Route::patch('/{member}', 'ResearchUserController@acceptMember');
-        Route::post('/decline/{member}', 'ResearchUserController@declineMember');
+        Route::patch('/{member}', 'ResearchUserController@acceptMember')->middleware('can:basic,member');
+        Route::post('/decline/{member}', 'ResearchUserController@declineMember')->middleware('can:basic,member');
         Route::post('/role/{member}', 'ResearchUserController@changeRole');
         Route::get('/role/detail/{member}', 'ResearchUserController@RoleUser');
     });
@@ -187,18 +187,18 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/', 'AcademicDegreeController@index');
         Route::get('/select-list', 'AcademicDegreeController@selectList');
         Route::get('/{academicDegree}', 'AcademicDegreeController@show');
-        Route::post('/', 'AcademicDegreeController@store');
-        Route::put('/{academicDegree}', 'AcademicDegreeController@store');
-        Route::delete('/{academicDegree}', 'AcademicDegreeController@destroy');
+        Route::post('/', 'AcademicDegreeController@store')->middleware('can:admin-only');
+        Route::put('/{academicDegree}', 'AcademicDegreeController@store')->middleware('can:admin-only');
+        Route::delete('/{academicDegree}', 'AcademicDegreeController@destroy')->middleware('can:admin-only');
     });
 
     Route::group(['prefix' => 'publication-type'], function () {
         Route::get('/', 'PublicationTypeController@index');
         Route::get('/select-list', 'PublicationTypeController@selectList');
         Route::get('/{publicationType}', 'PublicationTypeController@show');
-        Route::post('/', 'PublicationTypeController@store');
-        Route::put('/{publicationType}', 'PublicationTypeController@store');
-        Route::delete('/{publicationType}', 'PublicationTypeController@destroy');
+        Route::post('/', 'PublicationTypeController@store')->middleware('can:admin-only');
+        Route::put('/{publicationType}', 'PublicationTypeController@store')->middleware('can:admin-only');
+        Route::delete('/{publicationType}', 'PublicationTypeController@destroy')->middleware('can:admin-only');
     });
 
     Route::get('email/resend', 'VerificationController@resend')->name('verification.resend');
@@ -210,8 +210,8 @@ Route::group(['middleware' => ['jwt.verify', 'payment.status']], function () {
         Route::get('/{invoice}/show-file', 'PaymentController@showFile');
         Route::post('/{invoice}', 'PaymentController@storePayment');
         Route::post('/upload/{invoice}', 'PaymentController@storeUploadPayment');
-        Route::patch('/{invoice}/approve', 'PaymentController@acceptPayment');
-        Route::patch('/{invoice}/reject', 'PaymentController@rejectPayment');
+        Route::patch('/{invoice}/approve', 'PaymentController@acceptPayment')->middleware('can:basic,invoice');
+        Route::patch('/{invoice}/reject', 'PaymentController@rejectPayment')->middleware('can:basic,invoice');
     });
